requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-composed-permissions==0.2.1
drf-extra-fields==3.7.0
Faker==19.6.1
filetype==1.2.0
Markdown==3.4.4
Pillow==10.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
